% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tashu}
\alias{tashu}
\title{Bike rental history on public bicycle system("Tashu") from 2013 to 2015.}
\format{A data frame with 3394159 rows and 4 variables:
\describe{
  \item{RENT_STATION}{station Number where bike was rented.}
  \item{RENT_DATE}{date time when bike was rented. yyyy-mm-dd hh:mm:ss}
  \item{RETURN_STATION}{station Number where bike was returned.}
  \item{RETURN_DATE}{date time when bike was returned. yyyy-mm-dd hh:mm:ss}
}}
\usage{
tashu
}
\description{
A dataset which is bike rental history(rent station, rent date, return station, return date).
}
\keyword{datasets}
