% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weather}
\alias{weather}
\title{weather data in Daejeon, South Korea, from 2013 to 2015}
\format{A data frame with 26280 rows and 18 variables:
\describe{
  \item{Datetime}{date time}
  \item{Temperature}{Temperature}
  \item{Rainfall}{Rainfall}
  \item{WindSpeed}{Wind Speed(m/s)}
  \item{WindDirection}{Wind Direction}
  \item{Humidity}{Humidity}
  \item{VaporPressure}{Vapor Pressure(hpa)}
  \item{DewpointTemperature}{Dew point temperature(°C)}
  \item{LocalPressure}{Local Pressure(hpa)}
  \item{SeaSurfacePressure}{Sea Surface Pressure(hpa)}
  \item{Sunshine}{sun shine(hr)}
  \item{SolarRadiation}{solar radiation(MJ/m2)}
  \item{Snowfall}{snow fall(cm)}
  \item{GroundTemperature}{ground temperature(°C)}
  \item{X5cmGroundTemperature}{ground temperature(°C)}
  \item{X10cmGroundTemperature}{ground temperature(°C)}
  \item{X20cmGroundTemperature}{ground temperature(°C)}
  \item{X30cmGroundTemperature}{ground temperature(°C)}
}}
\usage{
weather
}
\description{
A dataset containing hourly weather data in Daejeon, 2013 ~ 2015
}
\keyword{datasets}
